+++ src/Exceptions/OllamaStreamException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace HalilCosdu\Ollama\Exceptions;
+
+use RuntimeException;
+
+class OllamaStreamException extends RuntimeException {}